modules/tensorrt_demo/face_alignment/requirements.txt
Pillow==6.0.0
numpy==1.16.3
opencv-python==3.4.1